local change

+++ src/Practice1.java
@@ -7,7 +7,7 @@ public class Practice1 {
         System.out.println(a + b);
         System.out.println(a - b);
         System.out.println(a * b);
-        System.out.println(a / b);
+        System.out.println((double)a / b);
         
         //This is the changes from github
     }
